refactor: cognitive-game 결과 저장 구조 정리 (routine update 미반영)
