+++ src/main/java/com/rememberme/dunoesanchaeg/contents/controller/CognitiveGameController.java
@@ -5,6 +5,7 @@ import com.rememberme.dunoesanchaeg.common.security.SecurityUtil;
 import com.rememberme.dunoesanchaeg.contents.dto.request.AnswerSubmitRequest;
 import com.rememberme.dunoesanchaeg.contents.dto.response.TodayGameResponse;
 import com.rememberme.dunoesanchaeg.contents.service.CognitiveGameService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -30,7 +31,7 @@ public class CognitiveGameController {
 
     // 2. 게임 결과 저장
     @PostMapping("/result")
-    public ApiResponse<Void> saveGameResult(@RequestBody AnswerSubmitRequest request) {
+    public ApiResponse<Void> saveGameResult(@Valid @RequestBody AnswerSubmitRequest request) {
         Long memberId = SecurityUtil.getCurrentMemberId();
 
         cognitiveGameService.saveGameResult(memberId, request);
