+++ src/main/java/com/rememberme/dunoesanchaeg/contents/dto/response/TodayGameResponse.java
@@ -1,14 +1,22 @@
 package com.rememberme.dunoesanchaeg.contents.dto.response;
 
 import com.rememberme.dunoesanchaeg.contents.domain.enums.GameType;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class TodayGameResponse {
     private LocalDate playedDate;
     private GameType gameType;
     private int totalRounds;
     private int roundTimeLimitSec;
     private String passCondition;
-    private boolean isGameFinished;
+    private Boolean isGameFinished;
 }
