+++ src/main/java/com/rememberme/dunoesanchaeg/contents/service/CognitiveGameService.java
@@ -1,48 +1,9 @@
 package com.rememberme.dunoesanchaeg.contents.service;
 
+import com.rememberme.dunoesanchaeg.contents.dto.request.AnswerSubmitRequest;
 import com.rememberme.dunoesanchaeg.contents.dto.response.TodayGameResponse;
-import com.rememberme.dunoesanchaeg.contents.mapper.CognitiveGameMapper;
-import com.rememberme.dunoesanchaeg.routines.domain.DailyRoutineStatus;
-import com.rememberme.dunoesanchaeg.routines.mapper.RoutineMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
-
-@Service
-@RequiredArgsConstructor
-public class CognitiveGameServiceImpl implements CognitiveGameService {
-
-    private final RoutineMapper routineMapper;
-    private final CognitiveGameMapper cognitiveGameMapper;
-
-    @Override
-    public TodayGameResponse getTodayGame(Long memberId) {
-
-        DailyRoutineStatus routine =
-                routineMapper.findByMemberIdAndDate(memberId, LocalDate.now());
-
-        if (routine == null) {
-            throw new IllegalArgumentException("오늘 게임 없음");
-        }
-
-        return TodayGameResponse.builder()
-                .gameType(routine.getAssignedGameType())
-                .isGameFinished(routine.getIsGameFinished())
-                .build();
-    }
-
-    @Override
-    public void saveGameResult(Long memberId, GameResultSaveRequest request) {
-
-        cognitiveGameMapper.insertGameLog(
-                memberId,
-                request.getGameType(),
-                request.getCorrectCount(),
-                request.getTotalTryCount()
-        );
-
-        // 루틴 상태 업데이트
-        routineMapper.updateGameFinished(memberId, LocalDate.now());
-    }
+public interface CognitiveGameService {
+    TodayGameResponse getTodayGame(Long memberId);
+    void saveGameResult(Long memberId, AnswerSubmitRequest request);
 }
