+++ src/main/resources/mapper/RoutineMapper.xml
@@ -44,4 +44,5 @@
                      #{createdAt}
                  );
     </insert>
+
 </mapper>
